Added in 3D Electric Ball Enemy and 3D Regular Ball Enemy
Imported both of the models Mohit made into a separate file called "3D Assets", Applied textures, rescaled model for unity and applied basic components to the models. To use them I've made a prefab for them simply drag them from the prefab folder onto the scene. Additionally they need lighting to show the textures so I added a light for 3D objects just throw that in the scene from the prefab folder as well and it should work.