(In Old Version) Particle System Implemented
Added in a particle system for when characters receive damage. The texture it uses is from the internet and will need to be replaced. If adding onto a new character attach the particle system to the character in the inspector. Currently already attached to GUNNER & SENTURY characters. This code can also easily be transferred to the new code if needed.